refs #3: Moved tests into django_tables dir for easier testing.
