+++ django_tables/tests/test_basic.py
@@ -7,6 +7,26 @@ from django.http import Http404
 from django.core.paginator import Paginator
 import django_tables as tables
 from django_tables.base import BaseTable
+from django.conf import settings
+
+
+def setup_module(module):
+    try:
+        local_settings = {
+            'DATABASES': {
+                'default': {
+                    'ENGINE': 'django.db.backends.sqlite3',
+                    'NAME': ':memory:',
+                },
+            },
+            'INSTALLED_APPS': ('tests.testapp',)
+        }
+        settings.configure(**local_settings)
+    except Exception:
+        pass
+
+    import django
+    django.setup()
 
 
 class TestTable(BaseTable):
