+++ django_tables/tests/test_memory.py
