+++ django_tables/tests/test_models.py
@@ -7,56 +7,10 @@ from nose.tools import assert_raises, assert_equal
 from django.conf import settings
 from django.core.paginator import Paginator, QuerySetPaginator
 import django_tables as tables
+from django_tables.tests.testapp.models import City, Country
 
 
 def setup_module(module):
-    local_settings = {
-        'DATABASES': {
-            'default': {
-                'ENGINE': 'django.db.backends.sqlite3',
-                'NAME': ':memory:',
-            },
-        },
-        'INSTALLED_APPS': ('tests.testapp',)
-    }
-    settings.configure(**local_settings)
-
-    from django.db import models
-    from django.core.management import call_command
-
-    class City(models.Model):
-        name = models.TextField()
-        population = models.IntegerField(null=True)
-
-        class Meta:
-            app_label = 'testapp'
-    module.City = City
-
-    class Country(models.Model):
-        name = models.TextField()
-        population = models.IntegerField()
-        capital = models.ForeignKey(City, blank=True, null=True)
-        tld = models.TextField(verbose_name='Domain Extension', max_length=2)
-        system = models.TextField(blank=True, null=True)
-        # tests expect this to be always null!
-        null = models.TextField(blank=True, null=True)
-        null2 = models.TextField(blank=True, null=True)  # - " -
-
-        def example_domain(self):
-            return 'example.%s' % self.tld
-
-        class Meta:
-            app_label = 'testapp'
-            db_table = 'country'
-    module.Country = Country
-
-    import django
-    if hasattr(django, 'setup'):
-        django.setup()
-
-    # create the tables
-    call_command('syncdb', verbosity=1, interactive=False)
-
     # create a couple of objects
     berlin = City.objects.create(name="Berlin", population=30)
     amsterdam = City.objects.create(name="Amsterdam", population=6)
@@ -81,6 +35,8 @@ class TestDeclaration:
     """
 
     def test_autogen_basic(self):
+        from django_tables.tests.testapp.models import City, Country
+
         class CountryTable(tables.ModelTable):
             class Meta:
                 model = Country  # noqa
@@ -139,6 +95,8 @@ class TestDeclaration:
     def test_columns_custom_order(self):
         """Using the columns meta option, you can also modify the ordering.
         """
+        from django_tables.tests.testapp.models import Country
+
         class CountryTable(tables.ModelTable):
             foo = tables.Column()
 
@@ -151,6 +109,8 @@ class TestDeclaration:
     def test_columns_verbose_name(self):
         """Tests that the model field's verbose_name is used for the column
         """
+        from django_tables.tests.testapp.models import Country
+
         class CountryTable(tables.ModelTable):
             class Meta:
                 model = Country  # noqa
@@ -189,6 +149,7 @@ def test_basic():
     Some tests here are copied from ``test_basic.py`` but need to be
     rerun with a ModelTable, as the implementation is different.
     """
+    from django_tables.tests.testapp.models import Country
 
     class CountryTable(tables.ModelTable):
         null = tables.Column(default="foo")
@@ -219,6 +180,8 @@ def test_basic():
 
 
 def test_with_filter():
+    from django_tables.tests.testapp.models import Country
+
     class CountryTable(tables.ModelTable):
         null = tables.Column(default="foo")
         domain = tables.Column(model_rel="tld")
@@ -236,6 +199,8 @@ def test_with_filter():
 
 
 def test_with_empty_list():
+    from django_tables.tests.testapp.models import Country
+
     class CountryTable(tables.ModelTable):
         null = tables.Column(default="foo")
         domain = tables.Column(model_rel="tld")
@@ -250,6 +215,8 @@ def test_with_empty_list():
 
 
 def test_with_no_results_query():
+    from django_tables.tests.testapp.models import Country
+
     class CountryTable(tables.ModelTable):
         null = tables.Column(default="foo")
         domain = tables.Column(model_rel="tld")
@@ -272,6 +239,8 @@ def test_invalid_accessor():
 
     Regression-Test: There used to be a NameError here.
     """
+    from django_tables.tests.testapp.models import Country
+
     class CountryTable(tables.ModelTable):
         name = tables.Column(model_rel='something-i-made-up')
     countries = CountryTable(Country)  # noqa
@@ -279,6 +248,8 @@ def test_invalid_accessor():
 
 
 def test_sort():
+    from django_tables.tests.testapp.models import City, Country
+
     class CountryTable(tables.ModelTable):
         domain = tables.Column(model_rel="tld")
         population = tables.Column()
@@ -332,6 +303,8 @@ def test_sort():
 
 
 def test_default_sort():
+    from django_tables.tests.testapp.models import Country
+
     class SortedCountryTable(tables.ModelTable):
         class Meta:
             model = Country  # noqa
@@ -360,6 +333,7 @@ def test_callable():
     """Some of the callable code is reimplemented for modeltables, so
     test some specifics again.
     """
+    from django_tables.tests.testapp.models import Country
 
     class CountryTable(tables.ModelTable):
         null = tables.Column(default=lambda s: s['example_domain'])
@@ -386,6 +360,7 @@ def test_callable():
 
 def test_relationships():
     """Test relationship spanning."""
+    from django_tables.tests.testapp.models import Country
 
     class CountryTable(tables.ModelTable):
         # add relationship spanning columns (using different approaches)
@@ -428,6 +403,7 @@ def test_pagination():
     Note: This test changes the available cities, make sure it is last,
     or that tests that follow are written appropriately.
     """
+    from django_tables.tests.testapp.models import City
     from django.db import connection
 
     class CityTable(tables.ModelTable):
@@ -479,6 +455,7 @@ def test_pagination():
 def test_evaluate_query():
     # We do not want queries passed in to be evaluated, we want to wait till
     # they are paginated
+    from django_tables.tests.testapp.models import Country
     from django.db import connection
 
     class CountryTable(tables.ModelTable):
@@ -500,6 +477,8 @@ def test_evaluate_query():
 
 
 def test_with_a_list():
+    from django_tables.tests.testapp.models import Country
+
     class CountryTable(tables.ModelTable):
         # add relationship spanning columns (using different approaches)
