--- tests/testapp/models.py
@@ -1,1 +1,0 @@
-"""Empty demo app our tests can assign models to."""
